Added AoE circle support for abilities and moved equipment classes to its own package.
