+++ src/dev/ipsych0/itemmaker/JSONWriter.java
@@ -1,7 +1,7 @@
 package dev.ipsych0.itemmaker;
 
 import com.google.gson.Gson;
-import dev.ipsych0.myrinnia.items.EquipSlot;
+import dev.ipsych0.myrinnia.equipment.EquipSlot;
 import dev.ipsych0.myrinnia.items.ItemRarity;
 import dev.ipsych0.myrinnia.items.ItemRequirement;
 import dev.ipsych0.myrinnia.items.ItemType;
