+++ src/dev/ipsych0/myrinnia/entities/creatures/Player.java
@@ -13,8 +13,8 @@ import dev.ipsych0.myrinnia.entities.npcs.ChatWindow;
 import dev.ipsych0.myrinnia.entities.npcs.ShopKeeper;
 import dev.ipsych0.myrinnia.gfx.Animation;
 import dev.ipsych0.myrinnia.gfx.Assets;
-import dev.ipsych0.myrinnia.items.EquipSlot;
-import dev.ipsych0.myrinnia.items.EquipmentWindow;
+import dev.ipsych0.myrinnia.equipment.EquipSlot;
+import dev.ipsych0.myrinnia.equipment.EquipmentWindow;
 import dev.ipsych0.myrinnia.items.InventoryWindow;
 import dev.ipsych0.myrinnia.items.ItemType;
 import dev.ipsych0.myrinnia.quests.QuestHelpUI;
